fix a bug introduced by bumping torch version.

+++ steganogan/models.py
@@ -323,7 +323,7 @@ class SteganoGAN(object):
 
         # split and decode messages
         candidates = Counter()
-        bits = image.data.cpu().numpy().tolist()
+        bits = image.data.int().cpu().numpy().tolist()
         for candidate in bits_to_bytearray(bits).split(b'\x00\x00\x00\x00'):
             candidate = bytearray_to_text(bytearray(candidate))
             if candidate:
